+++ R/verb-fill.R
@@ -70,18 +70,6 @@ fill.tbl_lazy <- function(.data, ..., .direction = c("down", "up", "updown", "do
   )
 }
 
-swap_order_direction <- function(x) {
-  if (is_quosure(x)) {
-    x <- quo_get_expr(x)
-  }
-
-  if (is_call(x, "desc", n = 1)) {
-    call_args(x)[[1]]
-  } else {
-    expr(desc(!!x))
-  }
-}
-
 dbplyr_fill0 <- function(.con, .data, cols_to_fill, order_by_cols, .direction) {
   UseMethod("dbplyr_fill0")
 }